Backend/requirements.txt
Django==5.2.7
python-dotenv
openai
djangorestframework
djangorestframework-simplejwt
pydantic
Pillow
